refactor: 지하철 역 데이터 생성 코드 리팩토링 - 역 좌표 정보 open api 관련 코드를 database - openapi 패키지로 이동
