+++ src/main/java/com/swyp/mema/database/openapi/location/converter/StationConverter.java
@@ -1,4 +1,4 @@
-package com.swyp.mema.domain.station.converter;
+package com.swyp.mema.database.openapi.location.converter;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -9,9 +9,9 @@ import com.swyp.mema.domain.station.dto.response.nearSubway.NearSubwayBasicRespo
 import com.swyp.mema.domain.station.dto.response.nearSubway.NearSubwayResponse;
 import com.swyp.mema.domain.station.dto.response.nearSubway.TotalNearSubwayResponse;
 import com.swyp.mema.domain.station.dto.response.subwayInfo.SingleStationResponse;
-import com.swyp.mema.domain.station.dto.response.subwayMaster.SubwayMasterBasicResponse;
-import com.swyp.mema.domain.station.dto.response.subwayMaster.SubwayMasterResponse;
-import com.swyp.mema.domain.station.dto.response.subwayMaster.TotalSubwayMasterResponse;
+import com.swyp.mema.database.openapi.location.response.SubwayMasterBasicResponse;
+import com.swyp.mema.database.openapi.location.response.SubwayMasterResponse;
+import com.swyp.mema.database.openapi.location.response.TotalSubwayMasterResponse;
 import com.swyp.mema.domain.station.dto.response.subwayTime.SubwayTimeBasicResponse;
 import com.swyp.mema.domain.station.dto.response.subwayTime.TotalSubwayTimeResponse;
 import com.swyp.mema.domain.station.dto.response.subwayTime.SubwayTimeResponse;
