+++ src/main/java/com/swyp/mema/domain/station/service/NearStationService.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import com.swyp.mema.domain.station.converter.StationConverter;
+import com.swyp.mema.database.openapi.location.converter.StationConverter;
 import com.swyp.mema.domain.station.dto.response.nearSubway.NearSubwayBasicResponse;
 import com.swyp.mema.domain.station.dto.response.nearSubway.TotalNearSubwayResponse;
